Cleanup database after tag group scenarios
