+++ features/007_tag_groups.feature
@@ -21,6 +21,7 @@
 Feature: Tag groups
   I can use Tag groups to organize tags in different topical areas.
 
+  @cleanup-after-scenario
   Scenario: Creating a tag group
     Given I create the tag "Test"
     And I go to the list of tag groups
@@ -34,6 +35,7 @@ Feature: Tag groups
     And I open the edit dialog of the tag "Test"
     Then the tag group "New tag group" is listed in the tag group list
   
+  @cleanup-after-scenario 
   Scenario: Adding existing tags to a tag group
     Given I create the tag "Test1"
     And I create the tag "Test2"
